fixed members ordering using LinkedHashMap

+++ ruby/execution_plan_utilities.rb
@@ -2,13 +2,15 @@ module ExecutionPlanUtilities
 
   EXECUTION_PLAN_FILE_NAME = "execution_plan.txt"
 
-  @counter = 0
+  FILES_PARENT = "C:\\Users\\Asus\\Documents\\Github\\pl-sql-compiler\\ruby\\"
+
+  @counter = 1
 
   def self.get_table_location(table_name)
 
     @json_array = JSON.parse File.read(MapReduce::DATA_TYPES_FILE_PATH)
     @json_array.map do |entry|
-      return entry["location"] if entry["name"].casecmp?(table_name)
+      return entry["location"], entry["field_terminator"] if entry["name"].casecmp?(table_name)
     end
   end
 
@@ -16,9 +18,11 @@ module ExecutionPlanUtilities
     Dir.entries(table_location).select {|f| (!File.directory? f) && (File.extname(f).casecmp?(".csv"))}
   end
 
-  def self.read_record(location, files, file_index, pos)
+  def self.read_record(location, files, file_index, pos, field_terminator)
 
-    line = nil
+    field_terminator = "\t" if field_terminator == "t"
+
+    line = ""
 
     File.open(location + "/" + files[file_index]) do |f|
 
@@ -35,7 +39,7 @@ module ExecutionPlanUtilities
 
     end
 
-    return line, file_index, pos
+    return line.gsub(/#{field_terminator}/, ","), file_index, pos
 
   end
 
@@ -109,7 +113,7 @@ module ExecutionPlanUtilities
 
   def self.init_execution_plan_file
 
-    File.delete(EXECUTION_PLAN_FILE_NAME)
+    File.delete(EXECUTION_PLAN_FILE_NAME) if File.exist?(EXECUTION_PLAN_FILE_NAME)
 
   end
 
@@ -120,6 +124,49 @@ module ExecutionPlanUtilities
     file.puts (@counter += 1).to_s + " - " + log
 
     file.close
+
+  end
+
+  def self.write_query_to_file(records, query_counter)
+
+    FileUtils.mkdir_p(FILES_PARENT + query_counter.to_s + "_query")
+
+    file = File.open(FILES_PARENT + query_counter.to_s + "_query/query.csv", "w")
+
+    file.puts records
+
+    file.close
+
+  end
+
+  def self.delete_query_files(query_counter)
+
+    query_counter.downto(0).each do |current_counter|
+
+      FileUtils.rm_rf(FILES_PARENT + current_counter.to_s + "_query")
+
+    end
+
+  end
+
+  def self.add_temp_data_type(name, members, query_counter)
+
+    temp_data_type = {}
+
+    temp_data_type["is_temp"] = true
+    temp_data_type["members"] = members
+    temp_data_type["name"] = name
+    temp_data_type["location"] = FILES_PARENT + query_counter.to_s
+
+    data_types = JSON.parse(File.read(MapReduce::DATA_TYPES_FILE_PATH))
+
+    data_types << temp_data_type
+
+    data_types_file = File.open(MapReduce::DATA_TYPES_FILE_PATH, "w")
+
+    data_types_file.puts data_types.to_json
+
+    data_types_file.close
 
   end
